requirements.txt
asgiref==3.5.0
Django==4.0.3
django-crispy-forms==1.14.0
djangorestframework==3.13.1
mysqlclient==2.1.0
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
